# server/openapi.yaml
openapi: 3.0.0
info:
  title: StellarYield API
  description: Backend API for the StellarYield yield farming and portfolio management platform
  version: 1.0.0
  contact:
    name: StellarYield Team
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.stellaryield.com
    description: Production server
...
paths:
...
  /api/yields:
    get:
      summary: Get current yield data
      description: Returns current yield information for all available vaults with APY, TVL, and risk ratings
      tags:
        - Yields
      responses:
        '200':
          description: Successful response with yields data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/YieldData'
        '500':
          description: Failed to fetch yield data
...
  /api/leaderboard:
    get:
      summary: Get user leaderboard
      description: Returns ranked list of users by TVL, yield earned, or other metrics
      tags:
        - Leaderboard
      parameters:
        - name: metric
          in: query
          schema:
            type: string
            enum: [tvl, yield, roi]
            default: tvl
          description: Metric to rank by
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [week, month, all]
            default: month
          description: Time period for ranking
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
      responses:
        '200':
          description: Leaderboard data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
                  total:
                    type: integer
        '400':
          description: Invalid parameters
...
  /api/alerts:
    post:
      summary: Create APY alert
      description: Creates a new APY alert for a wallet and vault
      tags:
        - Alerts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                  description: Stellar wallet address (G...)
                vaultId:
                  type: string
                  description: Vault identifier
                condition:
                  type: string
                  enum: [above, below]
                  description: Alert condition
                thresholdValue:
                  type: number
                  description: APY threshold value (0-1000)
                email:
                  type: string
                  format: email
                  description: Email address for notifications
              required:
                - walletAddress
                - vaultId
                - condition
                - thresholdValue
                - email
      responses:
        '201':
          description: Alert created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '400':
          description: Invalid alert parameters
        '429':
          description: Maximum alerts per user exceeded
        '500':
          description: Failed to create alert
...
  /api/auth/challenge:
    post:
      summary: Request authentication challenge
      description: Generates a challenge for wallet signing
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                  description: User's Stellar wallet address
      responses:
        '200':
          description: Challenge generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
                  walletAddressType:
                    type: string
                    enum: [account, contract]

    /metrics:
      get:
        summary: Prometheus scrape endpoint
        description: Prometheus-compatible metrics exposition. Returns plain-text metrics. This endpoint is protected by `METRICS_TOKEN` in production; unauthenticated requests return 404 to avoid token discovery.
        tags:
          - Metrics
        responses:
          '200':
            description: Prometheus metrics in text exposition format
            content:
              text/plain:
                schema:
                  type: string
          '404':
            description: Metrics endpoint not available or unauthorized
          '429':
            description: Too many requests (rate limited)
          '500':
            description: Internal server error while collecting metrics

    /api/audit-monitoring/alerts:
      get:
        summary: Get recent audit alerts
        description: Returns recent audit/monitoring alerts. Requires admin authentication.
        tags:
          - AuditMonitoring
        parameters:
          - name: severity
            in: query
            schema:
              type: string
              enum: [critical, warning, info]
          - name: type
            in: query
            schema:
              type: string
          - name: limit
            in: query
            schema:
              type: integer
              default: 100
        responses:
          '200':
            description: Alerts retrieved successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    count:
                      type: integer
                    alerts:
                      type: array
                      items:
                        $ref: '#/components/schemas/AuditAlert'
          '403':
            description: Admin access required
          '500':
            description: Failed to retrieve alerts

    /api/audit-monitoring/alerts/{severity}:
      get:
        summary: Get alerts filtered by severity
        description: Returns recent audit alerts for the given severity. Requires admin authentication.
        tags:
          - AuditMonitoring
        parameters:
          - name: severity
            in: path
            required: true
            schema:
              type: string
          - name: limit
            in: query
            schema:
              type: integer
              default: 100
        responses:
          '200':
            description: Alerts retrieved successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    count:
                      type: integer
                    severity:
                      type: string
                    alerts:
                      type: array
                      items:
                        $ref: '#/components/schemas/AuditAlert'
          '403':
            description: Admin access required
          '500':
            description: Failed to retrieve alerts

    /api/audit-monitoring/status:
      get:
        summary: Get monitoring status
        description: Returns current status of audit monitoring subsystem. Requires admin authentication.
        tags:
          - AuditMonitoring
        responses:
          '200':
            description: Monitoring status retrieved
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/MonitoringStatus'
          '403':
            description: Admin access required
          '500':
            description: Failed to retrieve monitoring status

    /api/audit-monitoring/report:
      get:
        summary: Generate a monitoring report
        description: Generates and returns a monitoring report (async generation may be implemented). Requires admin authentication.
        tags:
          - AuditMonitoring
        responses:
          '200':
            description: Monitoring report generated
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/MonitoringReport'
          '403':
            description: Admin access required
          '500':
            description: Failed to generate report

    /api/audit-monitoring/export:
      get:
        summary: Export monitoring data
        description: Returns a JSON export of monitoring data as an attachment. Requires admin authentication.
        tags:
          - AuditMonitoring
        responses:
          '200':
            description: Monitoring export (application/json attachment)
            content:
              application/json:
                schema:
                  type: object
          '403':
            description: Admin access required
          '500':
            description: Failed to export monitoring data

    /api/audit-monitoring/start:
      post:
        summary: Start audit monitoring
        description: Starts audit monitoring with provided configuration. Requires admin authentication.
        tags:
          - AuditMonitoring
        requestBody:
          required: false
          content:
            application/json:
              schema:
                type: object
        responses:
          '200':
            description: Monitoring started
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
                    status:
                      $ref: '#/components/schemas/MonitoringStatus'
          '403':
            description: Admin access required
          '500':
            description: Failed to start monitoring

    /api/audit-monitoring/stop:
      post:
        summary: Stop audit monitoring
        description: Stops audit monitoring. Requires admin authentication.
        tags:
          - AuditMonitoring
        responses:
          '200':
            description: Monitoring stopped
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
                    status:
                      $ref: '#/components/schemas/MonitoringStatus'
          '403':
            description: Admin access required
          '500':
            description: Failed to stop monitoring

  # ── Deposit Simulation ──────────────────────────────────────────────────────
...
  /api/contacts:
    get:
      summary: List all contacts
      description: |
        Returns all encrypted contacts for the authenticated wallet.
        Data is encrypted on the client side; the server stores and returns
        opaque base64 blobs without inspecting their contents.
      tags:
        - Contacts
      parameters:
        - name: x-wallet-address
          in: header
          required: true
          schema:
            type: string
          description: Stellar wallet address of the authenticated user
      responses:
        '200':
          description: Contact list returned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
                  total:
                    type: integer
        '401':
          description: Wallet address header missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactError'
    post:
      summary: Create a new contact
      tags:
        - Contacts
      parameters:
        - name: x-wallet-address
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - encryptedName
                - encryptedAddress
              properties:
                encryptedName:
                  type: string
                  description: Client-encrypted contact name (opaque base64 blob)
                encryptedAddress:
                  type: string
                  description: Client-encrypted Stellar address (opaque base64 blob)
      responses:
        '201':
          description: Contact created
          content:
            application/json:
              schema:
                type: object
                properties:
                  contact:
                    $ref: '#/components/schemas/Contact'
        '400':
          description: Missing or invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactError'
        '401':
          description: Wallet address header missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactError'
        '409':
          description: Contact with this encrypted address already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactError'
...
  /api/incidents:
    get:
      summary: List incidents
      description: Returns a paginated list of incidents ordered by start time descending.
      tags:
        - Incidents
      parameters:
        - name: protocol
          in: query
          schema:
            type: string
          description: Filter by protocol name
        - name: severity
          in: query
          schema:
            type: string
          description: Filter by severity level
        - name: type
          in: query
          schema:
            type: string
          description: Filter by incident type
        - name: resolved
          in: query
          schema:
            type: string
            enum: ["true", "false"]
          description: Filter by resolution status
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor from a previous response
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated incident list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Incident'
                  pagination:
                    type: object
                    properties:
                      nextCursor:
                        type: string
                        nullable: true
                      hasMore:
                        type: boolean
                      limit:
                        type: integer
        '500':
          description: Failed to fetch incidents
    post:
      summary: Create an incident
      tags:
        - Incidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - protocol
                - severity
                - type
                - title
                - description
                - startedAt
              properties:
                protocol:
                  type: string
                severity:
                  type: string
                type:
                  type: string
                title:
                  type: string
                description:
                  type: string
                affectedVaults:
                  type: array
                  items:
                    type: string
                startedAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Incident created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '400':
          description: Missing required fields
        '500':
          description: Failed to create incident
...
  /api/deposits/recommend:
    post:
      summary: Get deposit routing recommendation
      description: |
        Returns a routing recommendation for a multi-asset deposit: per-asset
        conversion/allocation path, expected vault-token output, estimated
        network fees, and warnings for unsupported assets.
      tags:
        - Deposits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - assets
              properties:
                assets:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - symbol
                      - amountInStroops
                    properties:
                      symbol:
                        type: string
                        description: Asset symbol (e.g. "USDC", "XLM")
                      amountInStroops:
                        type: string
                        description: Positive integer amount in stroops as a decimal string
      responses:
        '200':
          description: Routing recommendation
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid request body or asset parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
        '500':
          description: Failed to compute recommendation

  # Intentionally excluded routes:
  # - Internal debugging and profiling endpoints under /internal/* are intentionally omitted from this public spec.
  # - Any admin-only endpoints that require interactive session tokens are documented above when needed; ephemeral or highly-sensitive debug routes are excluded.
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        encrypted_name:
          type: string
          description: Client-encrypted contact name (opaque base64 blob)
        encrypted_address:
          type: string
          description: Client-encrypted Stellar address (opaque base64 blob)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - encrypted_name
        - encrypted_address
        - created_at
        - updated_at
    ContactError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
        - code

    # ── Incidents schemas ───────────────────────────────────────────────────
    Incident:
      type: object
      properties:
        id:
          type: string
        protocol:
          type: string
        severity:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        affectedVaults:
          type: array
          items:
            type: string
        startedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        resolved:
          type: boolean

    # ── Simulator schemas ───────────────────────────────────────────────────
...
    YieldData:
      type: object
      properties:
        protocol:
          type: string
          example: Blend
        asset:
          type: string
          example: USDC
        apy:
          type: number
          format: double
          example: 6.5
        tvl:
          type: number
          format: double
          example: 12000000
        risk:
          type: string
          enum: [Low, Medium, High]
        fetchedAt:
          type: string
          format: date-time
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        walletAddress:
          type: string
        tvl:
          type: number
          format: double
        yieldEarned:
          type: number
          format: double
        roi:
          type: number
          format: double
...
    Alert:
      type: object
      properties:
        id:
          type: string
        walletAddress:
          type: string
        vaultId:
          type: string
        condition:
          type: string
          enum: [above, below]
        thresholdValue:
          type: number
        email:
          type: string
          format: email
        isActive:
          type: boolean
        triggeredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    AuditAlert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        message:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
    MonitoringStatus:
      type: object
      properties:
        running:
          type: boolean
        startedAt:
          type: string
          format: date-time
          nullable: true
        lastCheckedAt:
          type: string
          format: date-time
          nullable: true
        alertCount:
          type: integer
        listeners:
          type: array
          items:
            type: string
    MonitoringReport:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
        summary:
          type: object
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/AuditAlert'
        metrics:
          type: object
        exportUrl:
          type: string
          description: URL to download a full export if available (admin-only)